backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
openai==1.3.0
python-dotenv==1.0.0
Pillow==10.1.0
transformers==4.35.0
torch==2.1.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.2
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
pymongo==4.6.1
cloudinary==1.37.0
motor==3.3.2
faiss-cpu==1.7.4 